backend/requirements.txt
flask==2.3.3
flask-restful==0.3.10
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
flask-migrate==4.0.5
pandas==2.1.1
scikit-learn==1.3.2
joblib==1.3.2
python-multipart==0.0.6
pytest==7.4.3
python-dotenv==1.0.0
sqlalchemy==2.0.23
mlflow
requests==2.31.0
beautifulsoup4==4.12.2 